requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.3.1
    # via
    #   -r requirements.in
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements.in
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   django-allauth
    #   django-cors-headers
astral==3.2
    # via -r requirements.in
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
    #   service-identity
    #   twisted
autobahn==24.4.2
    # via daphne
automat==24.8.1
    # via twisted
billiard==4.2.2
    # via
    #   -r requirements.in
    #   celery
celery==5.5.3
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2025.10.5
    # via
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
channels==4.1.0
    # via
    #   -r requirements.in
    #   channels-redis
    #   django-channels-jwt
channels-redis==4.1.0
    # via -r requirements.in
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via
    #   -r requirements.in
    #   celery
click-plugins==1.1.1.2
    # via
    #   -r requirements.in
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements.in
    #   celery
constantly==23.10.4
    # via twisted
cron-descriptor==1.4.5
    # via
    #   -r requirements.in
    #   django-celery-beat
cryptography==46.0.5
    # via
    #   -r requirements.in
    #   autobahn
    #   pyopenssl
    #   sendgrid
    #   service-identity
daphne==4.1.0
    # via -r requirements.in
defusedxml==0.7.1
    # via
    #   -r requirements.in
    #   python3-openid
dj-database-url==2.3.0
    # via -r requirements.in
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.2.12
    # via
    #   -r requirements.in
    #   channels
    #   dj-database-url
    #   dj-rest-auth
    #   django-allauth
    #   django-celery-beat
    #   django-cf-turnstile
    #   django-channels-jwt
    #   django-cors-headers
    #   django-extensions
    #   django-filter
    #   django-redis
    #   django-timezone-field
    #   django-vite
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-allauth==65.14.1
    # via -r requirements.in
django-celery-beat==2.8.1
    # via -r requirements.in
django-cf-turnstile==0.1.0
    # via -r requirements.in
django-channels-jwt==0.0.3
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
django-environ==0.12.0
    # via -r requirements.in
django-extensions==4.1
    # via -r requirements.in
django-filter==25.1
    # via -r requirements.in
django-ratelimit==4.1.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-timezone-field==7.1
    # via
    #   -r requirements.in
    #   django-celery-beat
django-vite==3.1.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
drf-spectacular==0.28.0
    # via -r requirements.in
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.25.1
    # via
    #   -r requirements.in
    #   drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
kombu==5.5.4
    # via
    #   -r requirements.in
    #   celery
markupsafe==3.0.3
    # via werkzeug
msgpack==1.1.2
    # via
    #   -r requirements.in
    #   channels-redis
numpy==2.3.4
    # via -r requirements.in
packaging==25.0
    # via kombu
prompt-toolkit==3.0.52
    # via click-repl
psycopg2==2.9.11
    # via -r requirements.in
psycopg2-binary==2.9.11
    # via -r requirements.in
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pyjwt==2.10.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
pyopenssl==26.0.0
    # via twisted
python-crontab==3.3.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
python-decouple==3.8
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
python-http-client==3.3.7
    # via sendgrid
python3-openid==3.2.0
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   drf-spectacular
redis==6.2.0
    # via
    #   -r requirements.in
    #   channels-redis
    #   django-redis
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   stripe
rpds-py==0.28.0
    # via
    #   jsonschema
    #   referencing
sendgrid==6.12.5
    # via -r requirements.in
sentry-sdk==2.43.0
    # via -r requirements.in
service-identity==24.2.0
    # via twisted
six==1.17.0
    # via python-dateutil
sqlparse==0.5.4
    # via
    #   -r requirements.in
    #   django
stripe==11.3.0
    # via -r requirements.in
twisted[tls]==24.11.0
    # via
    #   -r requirements.in
    #   daphne
txaio==23.1.1
    # via autobahn
typing-extensions==4.15.0
    # via
    #   -r requirements.in
    #   dj-database-url
    #   pyopenssl
    #   referencing
    #   stripe
    #   twisted
tzdata==2025.2
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   kombu
uritemplate==4.2.0
    # via
    #   -r requirements.in
    #   drf-spectacular
urllib3==2.6.3
    # via
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   -r requirements.in
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
werkzeug==3.1.6
    # via sendgrid
whitenoise==6.11.0
    # via -r requirements.in
zope-interface==7.2
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
